refactor(Fatras): Move particle selectors (#721)
Move all particle selector types into a single header in the Selectors to simplify the code structure. This ensures that the Fatras Kernel code does not contain any non-kernel code.

Remove the default selector types for the process wrapper helper. To ensure clear behavior, the user must provide. This was already the case everywhere in the examples anyway. Tests are updated accordingly.

* Fatras: combine particle selectors into single header

* Fatras: move EveryParticle to selectors

* Tests/Fatras: use mocks in process tests
